refactor(NumberUtil): 랜덤 숫자 생성을 Lotto 클래스에서 해주도록 변경
