+++ src/main/java/lotto/LottoNumber.java
@@ -1,8 +1,8 @@
 package lotto;
 
 public class LottoNumber implements Comparable<LottoNumber> {
-    private static final int MAX_LOTTO_NUMBER = 45;
-    private static final int MIN_LOTTO_NUMBER = 1;
+    static final int MAX_LOTTO_NUMBER = 45;
+    static final int MIN_LOTTO_NUMBER = 1;
     private final int number;
     private static final LottoNumber[] LIST = new LottoNumber[46];
